pyproject.toml
[project]
name = "claire"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.24",
    "pydantic-settings>=2.10.0",
    "python-gitlab>=6.0.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", # Missing docstrings
    "COM812", "ISC001", # Conflicting with formatter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "S101", # Allow assert statements in tests
    "PLR2004", # Allow magic values in tests
]
"examples/**/*.py" = [
    "T201", # Allow print statements in playground
    "INP001", # Allow implicit namespace packages in playground
]

[tool.ruff.lint.isort]
known-first-party = ["claire"]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
